🚚 refactor(endpoints): move route and rename solution commit endpoint to solutions instead of repository
