+++ models/solutions/solution.go
